+++ app/response_processor.py
@@ -21,6 +21,23 @@ class ResponseProcessor:
             pass
         return rv
 
+    def url_splitter(self, url=None):
+        if url is not None:
+            parts = url.split('/')
+
+            if 'responses' in parts:
+                service = 'responses'
+            elif 'decrypt' in parts:
+                service = 'decrypt'
+            elif 'validate' in parts:
+                service = 'validate'
+            else:
+                service = None
+        else:
+            service = None
+
+        return service
+
     def __init__(self, logger):
         self.logger = logger
         self.tx_id = ""
@@ -93,8 +110,10 @@ class ResponseProcessor:
         self.response_ok(self.remote_call(settings.SDX_RESPONSES_URL, json=decrypted_json))
 
     def remote_call(self, request_url, json=None, data=None, headers=None, verify=True, auth=None):
+        service = self.url_splitter(request_url)
+
         try:
-            self.logger.info("Calling service", request_url=request_url)
+            self.logger.info("Calling service", request_url=request_url, service=service)
             r = None
 
             if json:
@@ -110,17 +129,21 @@ class ResponseProcessor:
             self.logger.error("Max retries exceeded (5)", request_url=request_url)
 
     def response_ok(self, res):
+        request_url = res.url
+
+        service = self.url_splitter(request_url)
+
         res_logger = self.logger
         res_logger.bind(request_url=res.url, status=res.status_code)
 
         if res.status_code == 200 or res.status_code == 201:
-            res_logger.info("Returned from service", response="ok")
+            res_logger.info("Returned from service", response="ok", service=service)
             return
 
         elif res.status_code == 400:
-            res_logger.info("Returned from service", response="client error")
+            res_logger.info("Returned from service", response="client error", service=service)
             raise BadMessageError
 
         else:
-            res_logger.error("Returned from service", response="service error")
+            res_logger.error("Returned from service", response="service error", service=service)
             raise RetryableError